+++ README.md
@@ -4,41 +4,16 @@
 [![Test](https://github.com/devantler-tech/dotnet-age-cli/actions/workflows/test.yaml/badge.svg)](https://github.com/devantler-tech/dotnet-age-cli/actions/workflows/test.yaml)
 [![codecov](https://codecov.io/gh/devantler-tech/dotnet-age-cli/graph/badge.svg?token=RhQPb4fE7z)](https://codecov.io/gh/devantler-tech/dotnet-age-cli)
 
-<details>
-  <summary>Show/hide folder structure</summary>
-
-<!-- readme-tree start -->
-```
-.
-├── .github
-│   └── workflows
-├── scripts
-├── src
-│   └── Devantler.AgeCLI
-│       └── runtimes
-│           ├── linux-arm64
-│           │   └── native
-│           ├── linux-x64
-│           │   └── native
-│           ├── osx-arm64
-│           │   └── native
-│           ├── osx-x64
-│           │   └── native
-│           └── win-x64
-│               └── native
-└── tests
-    └── Devantler.AgeCLI.Tests
-        └── AgeKeygenTests
-
-20 directories
-```
-<!-- readme-tree end -->
-
-</details>
-
 A simple .NET library that embeds the Age CLI.
 
 ## 🚀 Getting Started
+
+### Prerequisites
+
+- .NET 9.0 or later
+- [Age CLI](https://github.com/FiloSottile/age#installation) installed and available in your system's PATH
+
+### Installation
 
 To get started, you can install the package from NuGet.
